Create separate Interface for each mapper, add to Koin module
