--- app/src/main/java/com/invillia/meubeats/data/mapper/EntityMapper.kt
@@ -1,12 +1,0 @@
-package com.invillia.meubeats.data.mapper
-
-interface EntityMapper<Entity, DomainModel> {
-
-    fun toDomain(entity: Entity): DomainModel
-
-    fun toEntity(domainModel: DomainModel): Entity
-
-    fun toDomainList(entityList: List<Entity>) = entityList.map { toDomain(it) }
-
-    fun toEntityList(domainList: List<DomainModel>) = domainList.map { toEntity(it) }
-}
